refactor: adjustments and refactoring: removed else statements, changed variable names for clarity
Refs #17
