+++ src/Autoload/Loader.php
@@ -13,19 +13,14 @@ declare(strict_types=1);
 
 namespace Phalcon\Autoload;
 
+use Phalcon\Autoload\Traits\LoaderHelperTrait;
+
 use function array_merge;
 use function array_unique;
 use function array_values;
-use function file_exists;
-use function is_array;
-use function is_string;
-use function rtrim;
 use function sha1;
 use function spl_autoload_register;
 use function spl_autoload_unregister;
-use function str_replace;
-use function strrpos;
-use function substr;
 use function trim;
 
 use const DIRECTORY_SEPARATOR;
@@ -34,45 +29,10 @@ use const DIRECTORY_SEPARATOR;
  * Class Logger
  *
  * @package Phalcon\Autoload
- *
- * @property array $classes
- * @property array $debug
- * @property array $extensions
- * @property array $files
- * @property bool  $isRegistered
- * @property array $namespaces
  */
 class Loader
 {
-    /**
-     * @var array
-     */
-    protected array $classes = [];
-
-    /**
-     * @var array
-     */
-    protected array $debug = [];
-
-    /**
-     * @var array
-     */
-    protected array $extensions = [];
-
-    /**
-     * @var array
-     */
-    protected array $files = [];
-
-    /**
-     * @var bool
-     */
-    protected bool $isRegistered = false;
-
-    /**
-     * @var array
-     */
-    protected array $namespaces = [];
+    use LoaderHelperTrait;
 
     /**
      * Loader constructor.
@@ -139,9 +99,9 @@ class Loader
         $directories,
         bool $prepend = false
     ): Loader {
-        $ns          = '\\';
-        $ds          = DIRECTORY_SEPARATOR;
-        $namespace   = trim($namespace, $ns) . $ns;
+        $nSeparator  = '\\';
+        $dSeparator  = DIRECTORY_SEPARATOR;
+        $namespace   = trim($namespace, $nSeparator) . $nSeparator;
         $directories = $this->checkDirectories($directories);
 
         // initialize the namespace prefix array if needed
@@ -149,7 +109,7 @@ class Loader
             $this->namespaces[$namespace] = [];
         }
 
-        $directories = $this->processDirectories($directories, $ds);
+        $directories = $this->processDirectories($directories, $dSeparator);
         $source      = ($prepend) ? $directories : $this->namespaces[$namespace];
         $target      = ($prepend) ? $this->namespaces[$namespace] : $directories;
 
@@ -356,140 +316,5 @@ class Loader
             spl_autoload_unregister([$this, 'autoload']);
             $this->isRegistered = false;
         }
-    }
-
-    /**
-     * Search for the file corresponding to the namespaced class and load it
-     *
-     * @param string $namespace
-     * @param string $class
-     *
-     * @return bool|string
-     */
-    protected function loadFile(string $namespace, string $class)
-    {
-        if (!isset($this->namespaces[$namespace])) {
-            $this->debug[] = 'Load: No folders registered: ' . $namespace;
-
-            return false;
-        }
-
-        return $this->processFileNameSpaces($namespace, $class);
-    }
-
-    /**
-     * If the file exists, require it and return true; false otherwise
-     *
-     * @param string $file The file to require
-     *
-     * @return bool
-     */
-    protected function requireFile(string $file): bool
-    {
-        if (file_exists($file)) {
-            require $file;
-
-            return true;
-        }
-
-        return false;
-    }
-
-    /**
-     * @param mixed $directories
-     *
-     * @return array
-     * @throws Exception
-     */
-    private function checkDirectories($directories): array
-    {
-        if (!is_string($directories) && !is_array($directories)) {
-            throw new Exception(
-                'The directories parameter is not a string or array'
-            );
-        }
-
-        if (is_string($directories)) {
-            $directories = [$directories];
-        }
-
-        return $directories;
-    }
-
-    /**
-     * @param array  $directories
-     * @param string $ds
-     *
-     * @return array
-     */
-    private function processDirectories(array $directories, string $ds): array
-    {
-        foreach ($directories as $key => $directory) {
-            $directories[$key] = rtrim($directory, $ds) . $ds;
-        }
-
-        return $directories;
-    }
-
-    /**
-     * @param string $namespace
-     * @param string $class
-     *
-     * @return bool|string
-     */
-    private function processFileNameSpaces(string $namespace, string $class)
-    {
-        $ns = '\\';
-        $ds = DIRECTORY_SEPARATOR;
-
-        foreach ($this->namespaces[$namespace] as $directory) {
-            foreach ($this->extensions as $extension) {
-                $file = $directory . str_replace($ns, $ds, $class) . '.' . $extension;
-
-                if ($this->requireFile($file)) {
-                    return $file;
-                }
-
-                $this->debug[] = 'Load: 404 : ' . $namespace . ' - ' . $file;
-            }
-        }
-
-        return false;
-    }
-
-    /**
-     * @param string $name
-     *
-     * @return bool|string
-     */
-    private function processLoadNameSpaces(string $name)
-    {
-        $ns = '\\';
-
-        /**
-         * Namespaces
-         *
-         * Matching in reverse the namespace names in relation to class names
-         */
-        $namespace = $name;
-        while (false !== $pos = strrpos($namespace, $ns)) {
-            // retain the trailing namespace separator in the prefix
-            $namespace = substr($name, 0, $pos + 1);
-            $remainder = substr($name, $pos + 1);
-
-            $file = $this->loadFile($namespace, $remainder);
-            if (false !== $file) {
-                $this->debug[] = 'Namespace: ' . $namespace . ' - ' . $file;
-
-                return $file;
-            }
-
-            $namespace = rtrim($namespace, $ns);
-        }
-
-        // 404
-        $this->debug[] = 'Namespace: 404 : ' . $name;
-
-        return false;
     }
 }
